Move AppDependencies to app/di package
Also renamed ModeSelectionUseCases to ModeSelectionDependencies for consistency.
